Updated the inventory system to use interactive UI
Created the DroppedItem script to handle the pick up / drop item logic.
Created the ItemUI script to handle single inventory UI slots.
Created the InventoryUI script to handle the interaction with the inventory.
Created the DragVisualTag script to handle the drag visual items.
Created the LimitedInventory script to consume the basic inventory script and having an inventory with limited slots.
Created the InventoryController script to show and hide the inventory.
Changed CursorController script location.