Added the save and write data system
Created the Data script, which will serve as a container for the data to be saved.
Created the DataReadWrite script to handle loading and saving data to a JSON file.
Created the LoadDroppedItems script to disable items that the player already has from the world and add them to the inventory.
Modified the LimitedRepository script to save data when adding, removing, and swapping items, as well as when the game ends.